+++ angular-view/src/app/service/server/url.service.ts
@@ -4,16 +4,19 @@ import { Injectable } from '@angular/core';
     providedIn: 'root'
 })
 export class UrlService {
-    readonly baseUrl: string;
+    //readonly baseUrl: string;
     readonly base: string;
     readonly auth_profile: string;
     readonly auth_google: string;
+    
+    readonly api_wordItems: string;
 
     constructor() {
-        this.baseUrl = window.location.origin + '/';
+        //this.baseUrl = window.location.origin + '/';
         this.base = '/';
         this.auth_profile = 'auth/profile';
         this.auth_google = 'auth/google';
+        this.api_wordItems = 'api/word-items';
     }
 
 }